Account Opening Form Created.

+++ src/main/java/com/example/banking_app/models/AddressModel.java
@@ -8,7 +8,7 @@ public class AddressModel extends BaseEntity{
 
     private String line1;
     private String line2;
-    private int zipCode;
+    private Long zipCode;
     private String city;
     private  String state;
     private String country;
@@ -32,11 +32,11 @@ public class AddressModel extends BaseEntity{
         this.line1 = line1;
     }
 
-    public int getZipCode() {
+    public Long getZipCode() {
         return zipCode;
     }
 
-    public void setZipCode(int zipCode) {
+    public void setZipCode(Long zipCode) {
         this.zipCode = zipCode;
     }
